extracted interfaces from serialization and deserialization services
