+++ src/SAREhub/Client/Event/EventSerializationService.php
@@ -1,15 +1,7 @@
 <?php
-
 namespace SAREhub\Client\Event;
 
-/**
- * Class used for serialize events
- */
-class EventSerializationService {
-	
-	/** @var callable[] */
-	private $serializerRegistry = [];
-	
+interface EventSerializationService {
 	/**
 	 * Register event serializer.
 	 * Serializer will be invoke with event object and must return array with all event data.
@@ -24,9 +16,7 @@ class EventSerializationService {
 	 * @param string $eventType
 	 * @param callable $serializer
 	 */
-	public function registerSerializer($eventType, callable $serializer) {
-		$this->serializerRegistry[$eventType] = $serializer;
-	}
+	public function registerSerializer($eventType, callable $serializer);
 	
 	/**
 	 * Serialize event object to json with serializer selected by event type
@@ -34,40 +24,17 @@ class EventSerializationService {
 	 * @return string
 	 * @throws EventSerializeException
 	 */
-	public function serialize(Event $event) {
-		if ($serializer = $this->getSerializer($event->getEventType())) {
-			if ($eventData = $serializer($event)) {
-				if ($json = json_encode($eventData, JSON_UNESCAPED_SLASHES | JSON_UNESCAPED_UNICODE)) {
-					return $json;
-				}
-				throw new EventSerializeException(
-				  "json_encode error: ".json_last_error_msg(), json_last_error()
-				);
-			}
-			throw new EventSerializeException(
-			  "Serializer must return array, given: ".var_export($eventData, true)
-			);
-		}
-		throw new EventSerializeException(
-		  "Serializer for event type: ".$event->getEventType()." isn't registered"
-		);
-	}
+	public function serialize(Event $event);
 	
 	/**
 	 * @param string $eventType
 	 * @return callable|null
 	 */
-	public function getSerializer($eventType) {
-		return $this->hasSerializer($eventType) ? $this->serializerRegistry[$eventType] : null;
-	}
+	public function getSerializer($eventType);
 	
 	/**
 	 * @param string $eventType
 	 * @return bool
 	 */
-	public function hasSerializer($eventType) {
-		return isset($this->serializerRegistry[$eventType]);
-	}
-	
-	
+	public function hasSerializer($eventType);
 }
