+++ tests/unit/SAREhub/Client/Event/JsonEventSerializationServiceTest.php
@@ -5,10 +5,10 @@ namespace SAREhub\Client\Event;
 use PHPUnit\Framework\TestCase;
 use SAREhub\Client\Event\User\UserEvent;
 
-class EventSerializationServiceTest extends TestCase {
+class JsonEventSerializationServiceTest extends TestCase {
 	
 	public function testRegisterSerializer() {
-		$serializationService = new EventSerializationService();
+		$serializationService = new JsonEventSerializationService();
 		
 		$serializerMock = $this->getCallbackMock();
 		$serializationService->registerSerializer("testEvent", $serializerMock);
@@ -37,7 +37,7 @@ class EventSerializationServiceTest extends TestCase {
 		$serializerMock = $this->getCallbackMock();
 		$serializerMock->expects($this->once())->method('__invoke')->with($eventMock)->willReturn($eventData);
 		
-		$serializationService = new EventSerializationService();
+		$serializationService = new JsonEventSerializationService();
 		$serializationService->registerSerializer("testEvent", $serializerMock);
 		$this->assertEquals(json_encode($eventData), $serializationService->serialize($eventMock));
 	}
@@ -48,7 +48,7 @@ class EventSerializationServiceTest extends TestCase {
 	public function testSerializeWithUnregisteredSerializer() {
 		$eventMock = $this->getMockBuilder(UserEvent::class)->disableOriginalConstructor()->getMock();
 		$eventMock->method('getEventType')->willReturn('testEvent');
-		$serializationService = new EventSerializationService();
+		$serializationService = new JsonEventSerializationService();
 		$serializationService->serialize($eventMock);
 	}
 	
@@ -72,7 +72,7 @@ class EventSerializationServiceTest extends TestCase {
 		$serializerMock = $this->getCallbackMock();
 		$serializerMock->expects($this->once())->method('__invoke')->with($eventMock)->willReturn(null);
 		
-		$serializationService = new EventSerializationService();
+		$serializationService = new JsonEventSerializationService();
 		$serializationService->registerSerializer("testEvent", $serializerMock);
 		$serializationService->serialize($eventMock);
 	}
